Updated and added support for various OSX packages.
Updated:
- 1password
- alfred
- caffeine
- ccmenu
- dropbox
- dmg_and_install script to install only packages that haven't been
  installed

Added:
- firefox
- iterm
- system_preferences.remote_desktop
- xquartz
- shiftit

+++ configurations/osx/states/osx_base/scripts/dmg_and_install.sh
@@ -1,10 +1,18 @@
 #!/bin/sh
-export dmg=$1
-export target_resource=$2
+dmg={{ pillar['caches']['packages'] }}$1
+target_resource=$2
+expected_sha=$3
+hashes_dir={{ pillar['caches']['hashes'] }}
+
+if [ "`cat "$dmg" | shasum -a 256 | cut -f 1 -d ' '`" == `cat "$hashes_dir/dmg_$target_resource"` ]; then
+    echo
+    echo "changed=no comment='$target_resource is already installed'"
+    exit 0
+fi
 
 rm -rf $destination || true
 volume=`hdiutil mount $dmg | grep Volumes | sed 's/.*\/Volumes/\/Volumes/g'`
 installer -pkg "$volume/$target_resource" -target /
 hdiutil detach "$volume"
 echo
-echo "changed=yes comment='$dmg >> $volume/$target_resource'"
+echo "changed=yes comment='installed $dmg >> $volume/$target_resource'"
